src/main/java/hu/nye/service/FinancialService.java
package hu.nye.service;

import hu.nye.model.User;
import hu.nye.model.Product;
import hu.nye.model.Expense;
import hu.nye.model.Invoice;
import hu.nye.model.Report;
import java.util.List;

/**
 * Service class for handling financial operations.
 * This class provides methods to generate invoices, track expenses, and generate financial reports.
 */
public class FinancialService implements FinancialServiceInterface {

    /**
     * Generates an invoice for the specified user and products.
     *
     * @param user The user for whom the invoice is generated.
     * @param products The list of products to include in the invoice.
     * @return The generated Invoice object.
     */
    @Override
    public Invoice generateInvoice(User user, List<Product> products) {
        // Invoice generation logic
        return new Invoice();
    }

    /**
     * Tracks a given expense.
     *
     * @param expense The expense to be tracked.
     */
    @Override
    public void trackExpense(Expense expense) {
        // Expense tracking logic
    }

    /**
     * Generates a financial report summarizing the financial data.
     *
     * @return The generated Report object.
     */
    @Override
    public Report generateFinancialReport() {
        // Financial report generation logic
        return new Report();
    }
}
